Add files via upload
I assumed the role of Data Analyst for my church, Lighthouse, starting in week four, focusing on tracking and analyzing membership growth. Our goal is to reach 200 members by the end of 2025, and we are identifying key lead and lag measures using scoreboards and data-driven insights to support strategic growth. Since I joined in week four as a new voluntary data analyst, I couldn't analyze and visualize data from weeks one to three due to time constraints, but I plan to work on them later. This project is dedicated to the growth of God's house, and I am fully committed to contributing my skills to its success.